# scan-engine/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "scan-to-map-studio API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/groups/{name}": {
      "get": {
        "summary": "Api Group",
        "operationId": "api_group_api_groups__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/prepare": {
      "post": {
        "summary": "Api Prepare",
        "operationId": "api_prepare_api_groups__name__prepare_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/upload": {
      "post": {
        "summary": "Api Upload Group",
        "operationId": "api_upload_group_api_groups_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_upload_group_api_groups_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupUploadResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/workspace": {
      "get": {
        "summary": "Api Workspace",
        "description": "Slices + alignment + metrics + floor images as JSON (Fleet Studio native workspace).",
        "operationId": "api_workspace_api_groups__name__workspace_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacePayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/groups/{name}/merged.slicemap.json": {
      "get": {
        "summary": "Api Merged Slicemap",
        "operationId": "api_merged_slicemap_api_groups__name__merged_slicemap_json_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlicemapV1"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/merged.floor.json": {
      "get": {
        "summary": "Api Merged Floor Json",
        "operationId": "api_merged_floor_json_api_groups__name__merged_floor_json_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FloorMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/alignment": {
      "get": {
        "summary": "Api Get Alignment",
        "operationId": "api_get_alignment_api_groups__name__alignment_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupAlignmentDoc"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Api Put Alignment",
        "operationId": "api_put_alignment_api_groups__name__alignment_put",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupAlignmentDoc"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveAlignmentResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/icp": {
      "post": {
        "summary": "Api Icp",
        "operationId": "api_icp_api_groups__name__icp_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PoseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IcpResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/merged.png": {
      "get": {
        "summary": "Api Merged Png",
        "operationId": "api_merged_png_api_groups__name__merged_png_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/groups/{name}/merged.floor.png": {
      "get": {
        "summary": "Api Merged Floor Png",
        "operationId": "api_merged_floor_png_api_groups__name__merged_floor_png_get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/groups/{name}": {
      "get": {
        "summary": "Group Workspace",
        "operationId": "group_workspace_groups__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_api_upload_group_api_groups_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_api_upload_group_api_groups_upload_post"
      },
      "FloorMeta": {
        "properties": {
          "format": {
            "type": "string",
            "title": "Format"
          },
          "resolution": {
            "type": "number",
            "title": "Resolution"
          },
          "origin": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Origin"
          },
          "width_px": {
            "type": "integer",
            "title": "Width Px"
          },
          "height_px": {
            "type": "integer",
            "title": "Height Px"
          },
          "row0": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row0"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "format",
          "resolution",
          "origin",
          "width_px",
          "height_px"
        ],
        "title": "FloorMeta",
        "description": "merged.floor.json (== studio.floorplan.floor_meta)."
      },
...
      "GroupAlignmentDoc": {
        "properties": {
          "format": {
            "type": "string",
            "title": "Format",
            "default": "scan-group-alignment-v1"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "up_axis_convention": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Up Axis Convention"
          },
          "alignments": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AlignmentEntry"
            },
            "type": "object",
            "title": "Alignments"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "reference",
          "alignments"
        ],
        "title": "GroupAlignmentDoc",
        "description": "group_alignment.json (scan-group-alignment-v1)."
      },
      "GroupStatus": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "dir": {
            "type": "string",
            "title": "Dir"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "scans": {
            "items": {
              "$ref": "#/components/schemas/ScanStatus"
            },
            "type": "array",
            "title": "Scans"
          },
          "has_alignment": {
            "type": "boolean",
            "title": "Has Alignment"
          },
          "has_merged": {
            "type": "boolean",
            "title": "Has Merged"
          },
          "ready": {
            "type": "boolean",
            "title": "Ready"
          },
          "has_floor": {
            "type": "boolean",
            "title": "Has Floor",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "dir",
          "reference",
          "scans",
          "has_alignment",
          "has_merged",
          "ready"
        ],
        "title": "GroupStatus"
      },
      "GroupUploadResult": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "scans": {
            "type": "integer",
            "title": "Scans"
          }
        },
        "type": "object",
        "required": [
          "status",
          "group",
          "url",
          "scans"
        ],
        "title": "GroupUploadResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IcpResult": {
        "properties": {
          "scan": {
            "type": "string",
            "title": "Scan"
          },
          "alignment": {
            "$ref": "#/components/schemas/Alignment"
          },
          "before": {
            "$ref": "#/components/schemas/AlignmentMetrics"
          },
          "after": {
            "$ref": "#/components/schemas/AlignmentMetrics"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/IcpStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "moved_m": {
            "type": "number",
            "title": "Moved M"
          },
          "rotated_deg": {
            "type": "number",
            "title": "Rotated Deg"
          }
        },
        "type": "object",
        "required": [
          "scan",
          "alignment",
          "before",
          "after",
          "steps",
          "moved_m",
          "rotated_deg"
        ],
        "title": "IcpResult"
      },
...
      "PoseRequest": {
        "properties": {
          "scan": {
            "type": "string",
            "title": "Scan"
          },
          "alignment": {
            "$ref": "#/components/schemas/Alignment"
          },
          "others": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/Alignment"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Others"
          }
        },
        "type": "object",
        "required": [
          "scan",
          "alignment"
        ],
        "title": "PoseRequest",
        "description": "Body of POST .../metrics and .../icp: one scan at a candidate pose, other scans' current poses."
      },
...
      "SaveAlignmentResult": {
        "properties": {
          "group": {
            "type": "string",
            "title": "Group"
          },
          "alignment_file": {
            "type": "string",
            "title": "Alignment File"
          },
          "merged": {
            "type": "string",
            "title": "Merged"
          },
          "merged_summary": {
            "type": "string",
            "title": "Merged Summary"
          },
          "cells": {
            "$ref": "#/components/schemas/MergedCells"
          },
          "scans": {
            "type": "integer",
            "title": "Scans"
          },
          "approved": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Approved"
          },
          "pending": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pending"
          },
          "published": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published"
          },
          "floor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floor"
          },
          "floor_scans": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Floor Scans",
            "default": []
          },
          "published_floor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Floor"
          }
        },
        "type": "object",
        "required": [
          "group",
          "alignment_file",
          "merged",
          "merged_summary",
          "cells",
          "scans",
          "approved",
          "pending"
        ],
        "title": "SaveAlignmentResult",
        "description": "PUT /api/groups/{name}/alignment -> merged slicemap rebuilt (+ published)."
      },
...
      "ScanStatus": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "has_folder": {
            "type": "boolean",
            "title": "Has Folder"
          },
          "has_usdz": {
            "type": "boolean",
            "title": "Has Usdz"
          },
          "has_project": {
            "type": "boolean",
            "title": "Has Project"
          },
          "has_slice": {
            "type": "boolean",
            "title": "Has Slice"
          },
          "method": {
            "type": "string",
            "title": "Method"
          }
        },
        "type": "object",
        "required": [
          "id",
          "has_folder",
          "has_usdz",
          "has_project",
          "has_slice",
          "method"
        ],
        "title": "ScanStatus"
      },
...
      "SlicemapV1": {
        "properties": {
          "format": {
            "type": "string",
            "const": "slicemap-v1",
            "title": "Format"
          },
          "z": {
            "type": "number",
            "title": "Z"
          },
          "band": {
            "type": "number",
            "title": "Band"
          },
          "resolution": {
            "type": "number",
            "title": "Resolution"
          },
          "origin": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Origin"
          },
          "cols": {
            "type": "integer",
            "title": "Cols"
          },
          "rows": {
            "type": "integer",
            "title": "Rows"
          },
          "data": {
            "type": "string",
            "title": "Data"
          },
          "sources": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SlicemapSource"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "format",
          "z",
          "band",
          "resolution",
          "origin",
          "cols",
          "rows",
          "data"
        ],
        "title": "SlicemapV1",
        "description": "slicemap-v1 file (merged.slicemap.json): base64 uint8 codes 0 unknown / 1 free / 2 furniture / 3 wall."
      },
...
      "WorkspacePayload": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/LayerPayload"
            },
            "type": "array",
            "title": "Layers"
          },
          "gates": {
            "$ref": "#/components/schemas/WorkspaceGates"
          },
          "api": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkspaceApi"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "group",
          "reference",
          "layers",
          "gates"
        ],
        "title": "WorkspacePayload",
        "description": "GET /api/groups/{name}/workspace (== studio.align_workspace_html.workspace_payload)."
      }
    }
  }
}